Fix crash on getting security policy when current v8::Context does not have an associated ScriptState. (#1008)
### Background and Motivation
A crash from Document::GetContentSecurityPolicyForWorld could happen when trying to move a loading CDC component. This indicate the current v8::Context doesn't have a blink::ScriptState associated with it. The problem only happen before the component is fully loaded.

### Description
Added a check for valid ScriptState.

### Test Plan
Run {CDC /I 7933099879<GO>}, and try to move it before it finish loading.

### Tickets
- [ENG2BBKIT-5103](https://jira.prod.bloomberg.com/browse/ENG2BBKIT-5103)
- {DRQS 164478933<go>}