Revert Scrollbar pseudo element style cache (#1029) (#1030)
This commit reverts the changeset
https://chromium.googlesource.com/chromium/src/+/05677fa5197561ebb83bbb4984696e75d44f466c%5E%21/

This changeset causes a soft memory leak, where the pseudo element style
cache within a ComputedStyle object keeps growing. In some cases, the
LayoutObject that is holding onto the ComputedStyle is GC'd and the
cache is freed. When this doesn't happen fast enough, the cache can
quickly consume 100s of MB over a few seconds and it can eventually
cause an OOM termination of the process.