Don't preallocate a render process host id (#1080)
Prior to this commit, we were preallocating a process host id when we
didn't want to impose a renderer affinity to a webview. This allocation
had a side effect that caused SiteInstanceImpl to query for the render
process host associated with the id and use it to host the webview.
This caused a problem because there were no process hosts associated
with the id and this caused the browser to crash.

This commit simply assigns kNoProcessAffinity as the affinity value.
This causes SiteInstanceImpl to skip over the attempt to pin down the
webview onto a specific renderer and let the upstream logic anchor the
webview onto any renderer.

This commit also moves the deletion of RequestContextManager to the
browser's IO thread, which is a requirement imposed by the destructor of
HeadlessRequestContextManager.