Focus webview window when it needs keyboard focus (#1090)
This commit reverts
https://source.chromium.org/chromium/chromium/src/+/5af4e86a.

The upstream change is valid only for the Chrome browser because all of
its input event processing is done for the top-level window. This is not
the case for blpwtk2 webviews so we must explicitly call SetFocus on the
webview window when we need to give it keyboard focus.